Update README.md
Several parts of this README look like a to-do list (apart from the actual TO-DO list from line 26 and following). Lines 14 and 20 indicate changes that should be done, and I don't know if this is the place for it. Also, on line 22 there is another feature that's not currently implemented (the relevant application clusters vs. all the clusters available).

In the CLI section, the `download` command should be added.

Apart from that, the edition includes grammar checking and rephrasing some sentences for clarity.